+++ src/react-components/src/components/01-atoms/Text/Headings/Heading.tsx
@@ -1,0 +1,52 @@
+// MT-82, MT 225, etc
+import * as React from "react";
+import bem from "../../../../utils/bem";
+
+export interface HeadingProps {
+  level: number;
+  id: string;
+  modifiers?: string[];
+  blockName?: string;
+
+  text?: string;
+  attributes?: {};
+}
+
+export default function Heading(props: React.PropsWithChildren<HeadingProps>) {
+  const { level, id, text, modifiers, blockName, attributes } = props;
+
+  if (level < 1 || level > 6) {
+    throw new Error("Heading only supports levels 1-6");
+  }
+
+  let content: string | React.ReactNode;
+  if (text) {
+    content = text;
+  } else {
+    let isValid: boolean = true;
+    React.Children.forEach(props.children, (child: React.ReactElement) => {
+      if (typeof child !== "string" && child.type !== "b") {
+        isValid = false;
+      }
+    });
+    if (!isValid) {
+      throw new Error("Headings can only be plain text or bold");
+    } else {
+      content = props.children;
+    }
+  }
+
+  let heading_base_class = "heading";
+
+  let headingProps = {
+    className: bem(heading_base_class, modifiers, blockName),
+    id: id,
+    ...attributes
+  };
+
+  return React.createElement(
+    "h" + level,
+    headingProps,
+    content
+  );
+}