+++ src/react-components/CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Currently, this repo is in Prerelease.  When it is released, this project will adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ========
+
+## PRERELEASE
+### Added
+- `Image` for 2:1 ratio
+- `PageTitle`
+- `Header with Image Right`
+
+### Changed 
+- Moved heading-related atoms from `01-atoms/Text` into `01-atoms/Text/Heading` 
+
 ## [0.0.5] - 2019-12-06
 ### Changed
 - `Icon` import